quarta alteração

+++ src/Telas/Shopping.jsx
@@ -1,79 +1,86 @@
-// Shopping.js
 import React, { useState, useEffect } from 'react';
 import { View, FlatList, StyleSheet, Text, Dimensions, TouchableOpacity, Alert } from 'react-native';
 import { Checkbox, TextInput, FAB, Appbar, List, PaperProvider } from 'react-native-paper';
-import { collection, addDoc, updateDoc, doc, deleteDoc, onSnapshot, query, where } from "firebase/firestore";
-import { db } from '../../services/firebaseConfig';
+import { collection, addDoc, updateDoc, doc, deleteDoc, onSnapshot, query, where, getDoc } from "firebase/firestore";
+import { db, auth } from '../../services/firebaseConfig';
 
 const { width } = Dimensions.get('window');
 
 export default function Shopping({ route, navigation }) {
-  const { store } = route.params; // Alteração no nome do parâmetro
+  const { store } = route.params;
   const [items, setItems] = useState([]);
   const [newItem, setNewItem] = useState('');
 
-  const EmojiCheckbox = ({ completed, onPress }) => {
-    return (
-      <TouchableOpacity onPress={onPress} accessible={true} accessibilityLabel={completed ? "Desmarcar item" : "Marcar item como concluído"}>
-        <Text style={{ fontSize: 24 }}>
-          {completed ? '✅' : '⬜'} {/* Alterna entre os emojis de checked e unchecked */}
-        </Text>
-      </TouchableOpacity>
-    );
-  };
+  const EmojiCheckbox = ({ completed, onPress }) => (
+    <TouchableOpacity style={{marginLeft:10}} onPress={onPress} accessible={true} accessibilityLabel={completed ? "Desmarcar item" : "Marcar item como concluído"}>
+      <Text style={{ fontSize: 24 }}>
+        {completed ? '✅' : '⬜'}
+      </Text>
+    </TouchableOpacity>
+  );
 
   useEffect(() => {
-    if (store && store.id) {
-      const itemsRef = collection(db, 'shoppingItems'); // Alteração para 'shoppingItems'
-      const q = query(itemsRef, where('storeId', '==', store.id));
+    const user = auth.currentUser;
+    if (user) {
+      const q = query(
+        collection(db, 'shoppingItems'),
+        where('storeId', '==', store.id),
+        where('userId', '==', user.uid)
+      );
       const unsubscribe = onSnapshot(q, (snapshot) => {
         const fetchedItems = snapshot.docs.map(doc => ({
           id: doc.id,
-          ...doc.data(),
+          ...doc.data()
         }));
         setItems(fetchedItems);
       });
 
       return () => unsubscribe();
     }
-  }, [store]);
+  }, [store.id]);
 
   const addItem = async () => {
     if (newItem.trim()) {
       try {
-        await addDoc(collection(db, 'shoppingItems'), { // Alteração para 'shoppingItems'
-          name: newItem,
-          completed: false,
-          storeId: store.id, // Referência da loja
-        });
-        setNewItem('');
+        const user = auth.currentUser;
+        if (user) {
+          await addDoc(collection(db, 'shoppingItems'), {
+            title: newItem,
+            completed: false,
+            storeId: store.id,
+            userId: user.uid,
+          });
+          setNewItem('');
+        }
       } catch (error) {
         console.error('Erro ao adicionar item de compra:', error);
         Alert.alert('Erro', 'Não foi possível adicionar o item. Tente novamente.');
       }
-    } else {
-      Alert.alert('Aviso', 'Por favor, insira o nome do item.');
     }
   };
 
-  const toggleItemCompletion = async (id, completed) => {
+  const toggleComplete = async (id, completed) => {
     try {
-      const itemRef = doc(db, 'shoppingItems', id); // Alteração para 'shoppingItems'
-      await updateDoc(itemRef, {
-        completed: !completed,
-      });
+      const itemRef = doc(db, 'shoppingItems', id);
+      await updateDoc(itemRef, { completed: !completed });
     } catch (error) {
-      console.error('Erro ao atualizar item:', error);
+      console.error('Erro ao atualizar item de compra:', error);
       Alert.alert('Erro', 'Não foi possível atualizar o item. Tente novamente.');
     }
   };
 
   const deleteItem = async (id) => {
     try {
-      const itemRef = doc(db, 'shoppingItems', id); // Alteração para 'shoppingItems'
-      await deleteDoc(itemRef);
+      const user = auth.currentUser;
+      const itemRef = doc(db, 'shoppingItems', id);
+      const itemDoc = await getDoc(itemRef);
+      if (itemDoc.exists() && itemDoc.data().userId === user.uid) {
+        await deleteDoc(itemRef);
+      } else {
+        Alert.alert('Erro', 'Você não tem permissão para deletar este item.');
+      }
     } catch (error) {
-      console.error('Erro ao deletar item:', error);
+      console.error('Erro ao deletar item de compra:', error);
       Alert.alert('Erro', 'Não foi possível deletar o item. Tente novamente.');
     }
   };
@@ -81,7 +88,7 @@ export default function Shopping({ route, navigation }) {
   return (
     <PaperProvider>
       <View style={styles.container}>
-        <Appbar.Header style={{ backgroundColor: '#6a6e73', borderWidth: 4, borderRadius: 10, paddingLeft:20 }}>
+        <Appbar.Header style={styles.appbar}>
           <Appbar.Content titleStyle={styles.appbarTitle} title={store ? store.title : 'Shopping List'} />
         </Appbar.Header>
 
@@ -90,23 +97,21 @@ export default function Shopping({ route, navigation }) {
           keyExtractor={(item) => item.id}
           renderItem={({ item }) => (
             <List.Item
-              title={item.name}
+              title={item.title}
               style={styles.listItem}
               titleStyle={[styles.titleStyle, item.completed && styles.completedTitle]}
               left={() => (
-                <View style={styles.leftContainer}>
-                  <EmojiCheckbox
-                    completed={item.completed}
-                    onPress={() => toggleItemCompletion(item.id, item.completed)}
-                  />
-                </View>
+                <EmojiCheckbox
+                  completed={item.completed}
+                  onPress={() => toggleComplete(item.id, item.completed)}
+                />
               )}
               right={() => (
                 <Text 
                   style={styles.delete} 
                   onPress={() => deleteItem(item.id)}
                   accessible={true}
-                  accessibilityLabel={`Deletar item ${item.name}`}
+                  accessibilityLabel={`Deletar item ${item.title}`}
                 >
                   🗑️
                 </Text>
@@ -136,7 +141,6 @@ export default function Shopping({ route, navigation }) {
           accessible={true}
           accessibilityLabel="Botão flutuante para adicionar um novo item de compra"
         />
-
       </View>
     </PaperProvider>
   );
@@ -146,11 +150,26 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     padding: 16,
-    backgroundColor: '#37474F',
+    backgroundColor: '#F5F5F5',
+  },
+  appbar: {
+    backgroundColor: '#3E4A59',
+    borderWidth: 4,
+    borderRadius: 10,
+    paddingLeft: 20,
+  },
+  appbarTitle: {
+    color: '#FFF',
+    fontStyle: 'italic',
+    fontWeight: 'bold',
+    fontSize: 25,
+    textShadowColor: 'rgba(0, 0, 0, 0.55)',
+    textShadowOffset: { width: -1, height: 1 },
+    textShadowRadius: 10,
   },
   listItem: {
     width: '100%',
-    backgroundColor: '#ECEFF1',
+    backgroundColor: '#F5F5DC',
     marginVertical: 4,
     marginTop: 20,
     borderRadius: 20,
@@ -159,7 +178,6 @@ const styles = StyleSheet.create({
     fontSize: 22,
     color: '#37474F',
     fontWeight: 'bold',
-    fontStyle: 'italic',
     textShadowColor: 'rgba(0, 0, 0, 0.45)',
     textShadowOffset: { width: -1, height: 1 },
     textShadowRadius: 10,
@@ -168,20 +186,11 @@ const styles = StyleSheet.create({
     textDecorationLine: 'line-through',
     color: 'gray',
   },
-  leftContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginLeft: 10,
-  },
-  icon: {
-    fontSize: 18,
-    marginRight: 8,
-  },
   emptyContainer: {
     flex: 1,
     justifyContent: 'center',
     alignItems: 'center',
-    marginTop:50
+    marginTop: 50,
   },
   emptyMessage: {
     fontSize: 18,
@@ -190,15 +199,16 @@ const styles = StyleSheet.create({
   input: {
     width: '80%',
     borderRadius: 7,
-    padding:10,
+    padding: 9 ,
     marginTop: 20,
+    backgroundColor: '#3E4A59',
   },
   fab: {
     position: 'absolute',
     margin: 16,
     right: 0,
     bottom: 0,
-    backgroundColor: '#FF5722',
+    backgroundColor: '#03A9F4',
     width: width * 0.15,
     height: width * 0.15,
     borderRadius: width * 0.075,
@@ -216,14 +226,5 @@ const styles = StyleSheet.create({
     padding: 8,
     color: 'red',
     fontSize: 18,
-  },
-  appbarTitle: {
-    color: '#FFF',
-    fontStyle: 'italic',
-    fontWeight: 'bold',
-    fontSize: 25,
-    textShadowColor: 'rgba(0, 0, 0, 0.55)',
-    textShadowOffset: { width: -1, height: 1 },
-    textShadowRadius: 10,
   },
 });
